refactor: handlers for catalog
